requirements.txt
Flask==0.8
discord.py[voice]==0.16.6
